refactor folder strucutres and Prop types
