+++ frontend-opal/components/PortfolioContainer/ValueInfo/index.js
@@ -1,34 +1,3 @@
-import React from 'react';
-import PropTypes from 'prop-types';
-import { View, Text, TouchableOpacity } from 'react-native';
-import { Card } from 'native-base';
-import styles from './styles';
-
-const ValueInfo = ({ totalInvestment }) => {
-    return (
-      <View style={styles.valueArea}>
-        <Text style={styles.title}>Value</Text>
-
-        <Card>
-          <View style={styles.investment}>
-            <Text style={styles.name}>Total Investment</Text>
-            <Text style={styles.amount}>
-
-              {totalInvestment}
-            </Text>
-          </View>
-
-          <TouchableOpacity onPress={() => console.log('button pressed')}>
-            <Text style={styles.button}>INVESTMENT ACCOUNT</Text>
-          </TouchableOpacity>
-
-        </Card>
-      </View>
-    )
-}
-
-ValueInfo.propTypes = {
-  totalInvestment: PropTypes.string.isRequired
-}
+import ValueInfo from './ValueInfo';
 
 export default ValueInfo;
