+++ frontend-opal/components/Portofolio.js
@@ -3,7 +3,7 @@ import { View } from 'react-native';
 import Header from './Header';
 import PortfolioContainer from './PortfolioContainer'
 
-export default function OpalApp() {
+export default function Portfolio() {
   return (
     <View>
       <Header />
